requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
colorama==0.4.6
pytest-crayons==0.0.5
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
more-itertools==9.1.0
packaging==23.1
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.3.1
requests==2.31.0
selenium==4.10.0
toml==0.10.2
urllib3==2.0.3
wcwidth==0.2.6
webdriver-manager==3.8.6





